+++ src/FabulousScheduler.Queue/Enums/QueueJobStateEnum.cs
@@ -1,6 +1,5 @@
 namespace FabulousScheduler.Queue.Enums;
 
-[Flags]
 public enum QueueJobStateEnum
 {
 	/// <summary>